renamed folder, changed readme
